grid, bar: move AppBox to nwg_classes.
